requirements.txt
discord~=1.7.3
Pillow~=8.3.2
python-dotenv~=0.19.0
nextcord~=2.0.0a2
mutagen~=1.45.1
aiohttp~=3.7.4.post0